rename the directory assistant/test to assistant/tests
fix #2163
